actions/embed_images_to_pinecone/requirements.dev.txt
# Python packages to install into the this directory's virtual environment
# for the purpose of development, testing, and deployment.
# pytorch-cpu is also required but installed separately
roboto
numpy
jupyter
matplotlib
pinecone-client
transformers
jupyter
